+++ examples/nlp/token_classification/punctuate_capitalize_infer.py
@@ -13,23 +13,10 @@
 # limitations under the License.
 
 import argparse
-import itertools
 import json
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
 
-import numpy as np
-import torch
-from omegaconf import DictConfig
-from torch.nn.utils.rnn import pad_sequence
-
-from nemo.collections.common.tokenizers.tokenizer_spec import TokenizerSpec
-from nemo.collections.nlp.data.data_utils.data_preprocessing import get_stats
 from nemo.collections.nlp.models import PunctuationCapitalizationModel
-from nemo.core.classes import Dataset
-from nemo.core.neural_types import ChannelType, Index, MaskType, NeuralType
-from nemo.core.neural_types.elements import BoolType
-from nemo.utils import logging
 
 
 """
@@ -170,536 +157,6 @@ def load_manifest(manifest: Path):
     return result
 
 
-def _get_subtokens_and_subtokens_mask(query: str, tokenizer: TokenizerSpec) -> Tuple[List[str], List[int]]:
-    """
-    Tokenizes input query into subtokens and creates subtokens mask. Subtokens mask is an array of the same length as
-    subtokens array and contains zeros and ones in which. If element of mask equals 1, then corresponding subtoken in
-    subtokens array is first subtoken in some word
-    Args:
-        query: a string that will be tokenized
-        tokenizer: an instance of tokenizer
-    Returns:
-        subtokens: list of subtokens
-        subtokens_mask: list of ints
-    """
-    words = query.strip().split()
-    subtokens = []
-    subtokens_mask = []
-    for j, word in enumerate(words):
-        word_tokens = tokenizer.text_to_tokens(word)
-        subtokens.extend(word_tokens)
-        subtokens_mask.append(1)
-        subtokens_mask.extend([0] * (len(word_tokens) - 1))
-    return subtokens, subtokens_mask
-
-
-def _check_max_seq_length_and_margin_and_step(max_seq_length: int, margin: int, step: int):
-    """
-    Checks values of ``max_seq_length``, ``margin``, and ``step``.
-    Args:
-        max_seq_length: a segment length with ``[CLS]`` and ``[SEP]`` tokens
-        margin: a number of input tokens near edges of segments which are not used in punctuation and capitalization
-            prediction.
-        step: offset of consequent segments.
-    Returns:
-        None
-    """
-    if max_seq_length < 3:
-        raise ValueError(
-            f"Parameter `max_seq_length={max_seq_length}` cannot be less than 3 because `max_seq_length` is a length "
-            f"of a segment with [CLS] and [SEP] tokens."
-        )
-    if margin >= (max_seq_length - 2) // 2 and margin > 0 or margin < 0:
-        raise ValueError(
-            f"Parameter `margin` has to be not negative and less than `(max_seq_length - 2) // 2`. Don't forget about "
-            f"CLS and EOS tokens in the beginning and the end of segment. margin={margin}, "
-            f"max_seq_length={max_seq_length}"
-        )
-    if step <= 0:
-        raise ValueError(f"Parameter `step` has to be positive whereas step={step}")
-    if step > max_seq_length - 2 - 2 * margin:
-        logging.warning(
-            f"Parameter step={step} is too big. It will be reduced to `min(max_seq_length, <maximum query length> + 2) "
-            f"- 2 - 2 * margin`."
-        )
-
-
-def get_features_infer(
-    queries: List[str],
-    tokenizer: TokenizerSpec,
-    max_seq_length: int = 64,
-    step: Optional[int] = 8,
-    margin: Optional[int] = 16,
-) -> Tuple[
-    List[List[int]], List[List[int]], List[List[int]], List[List[int]], List[int], List[int], List[bool], List[bool],
-]:
-    """
-    Processes the data and returns features.
-
-    Args:
-        queries: text sequences
-        tokenizer: such as AutoTokenizer
-        max_seq_length: max sequence length minus 2 for [CLS] and [SEP]
-        step: relative shift of consequent segments into which long queries are split. Long queries are split into
-            segments which can overlap. Parameter ``step`` controls such overlapping. Imagine that queries are
-            tokenized into characters, ``max_seq_length=5``, and ``step=2``. In such a case query "hello" is
-            tokenized into segments ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'l', 'l', 'o', '[SEP]']]``.
-        margin: number of subtokens near edges of segments which are not used for punctuation and capitalization
-            prediction. The first segment does not have left margin and the last segment does not have right
-            margin. For example, if input sequence is tokenized into characters, ``max_seq_length=5``,
-            ``step=1``, and ``margin=1``, then query "hello" will be tokenized into segments
-            ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'e', 'l', 'l', '[SEP]'],
-            ['[CLS]', 'l', 'l', 'o', '[SEP]']]``. These segments are passed to the model. Before final predictions
-            computation, margins are removed. In the next list, subtokens which logits are not used for final
-            predictions computation are marked with asterisk: ``[['[CLS]'*, 'h', 'e', 'l'*, '[SEP]'*],
-            ['[CLS]'*, 'e'*, 'l', 'l'*, '[SEP]'*], ['[CLS]'*, 'l'*, 'l', 'o', '[SEP]'*]]``.
-
-    Returns:
-        all_input_ids: list of input ids of all segments
-        all_segment_ids: token type ids of all segments
-        all_input_mask: attention mask to use for BERT model
-        all_subtokens_mask: masks out all subwords besides the first one
-        all_quantities_of_preceding_words: number of words in query preceding a segment. Used for joining
-            predictions from overlapping segments.
-        all_query_ids: index of a query to which segment belongs
-        all_is_first: is segment first segment in a query
-        all_is_last: is segment last segment in a query
-    """
-    st = []
-    stm = []
-    sent_lengths = []
-    for i, query in enumerate(queries):
-        subtokens, subtokens_mask = _get_subtokens_and_subtokens_mask(query, tokenizer)
-        sent_lengths.append(len(subtokens))
-        st.append(subtokens)
-        stm.append(subtokens_mask)
-    _check_max_seq_length_and_margin_and_step(max_seq_length, margin, step)
-    max_seq_length = min(max_seq_length, max(sent_lengths) + 2)
-    logging.info(f'Max length: {max_seq_length}')
-    # Maximum number of word subtokens in segment. The first and the last tokens in segment are CLS and EOS
-    length = max_seq_length - 2
-    step = min(length - margin * 2, step)
-    get_stats(sent_lengths)
-    all_input_ids, all_segment_ids, all_subtokens_mask, all_input_mask, all_input_mask = [], [], [], [], []
-    all_quantities_of_preceding_words, all_query_ids, all_is_first, all_is_last = [], [], [], []
-    for q_i, query_st in enumerate(st):
-        q_inp_ids, q_segment_ids, q_subtokens_mask, q_inp_mask, q_quantities_of_preceding_words = [], [], [], [], []
-        for i in range(0, max(len(query_st), length) - length + step, step):
-            subtokens = [tokenizer.cls_token] + query_st[i : i + length] + [tokenizer.sep_token]
-            q_inp_ids.append(tokenizer.tokens_to_ids(subtokens))
-            q_segment_ids.append([0] * len(subtokens))
-            q_subtokens_mask.append([0] + stm[q_i][i : i + length] + [0])
-            q_inp_mask.append([1] * len(subtokens))
-            q_quantities_of_preceding_words.append(np.count_nonzero(stm[q_i][:i]))
-        all_input_ids.append(q_inp_ids)
-        all_segment_ids.append(q_segment_ids)
-        all_subtokens_mask.append(q_subtokens_mask)
-        all_input_mask.append(q_inp_mask)
-        all_quantities_of_preceding_words.append(q_quantities_of_preceding_words)
-        all_query_ids.append([q_i] * len(q_inp_ids))
-        all_is_first.append([True] + [False] * (len(q_inp_ids) - 1))
-        all_is_last.append([False] * (len(q_inp_ids) - 1) + [True])
-    return (
-        list(itertools.chain(*all_input_ids)),
-        list(itertools.chain(*all_segment_ids)),
-        list(itertools.chain(*all_input_mask)),
-        list(itertools.chain(*all_subtokens_mask)),
-        list(itertools.chain(*all_quantities_of_preceding_words)),
-        list(itertools.chain(*all_query_ids)),
-        list(itertools.chain(*all_is_first)),
-        list(itertools.chain(*all_is_last)),
-    )
-
-
-class BertPunctuationCapitalizationInferLongDataset(Dataset):
-    """
-    Creates dataset to use during inference for punctuation and capitalization tasks with a pretrained model.
-    For dataset to use during training with labels, see BertPunctuationCapitalizationDataset.
-
-    Parameters ``max_seq_length``, ``step``, ``margin`` are for controlling the way queries are split into segments
-    which then processed by the model. Parameter ``max_seq_length`` is a length of a segment after tokenization
-    including special tokens [CLS] in the beginning and [SEP] in the end of a segment. Parameter ``step`` is shift
-    between consequent segments. Parameter ``margin`` is used to exclude negative effect of subtokens near
-    borders of segments which have only one side context.
-
-    Args:
-        queries: list of sequences.
-        tokenizer: such as AutoTokenizer
-        max_seq_length: max sequence length minus 2 for [CLS] and [SEP]
-        step: relative shift of consequent segments into which long queries are split. Long queries are split into
-            segments which can overlap. Parameter ``step`` controls such overlapping. Imagine that queries are
-            tokenized into characters, ``max_seq_length=5``, and ``step=2``. In such a case query "hello" is
-            tokenized into segments ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'l', 'l', 'o', '[SEP]']]``.
-        margin: number of subtokens in the beginning and the end of segments which are not used for prediction
-            computation. The first segment does not have left margin and the last segment does not have right
-            margin. For example, if input sequence is tokenized into characters, ``max_seq_length=5``,
-            ``step=1``, and ``margin=1``, then query "hello" will be tokenized into segments
-            ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'e', 'l', 'l', '[SEP]'],
-            ['[CLS]', 'l', 'l', 'o', '[SEP]']]``. These segments are passed to the model. Before final predictions
-            computation, margins are removed. In the next list, subtokens which logits are not used for final
-            predictions computation are marked with asterisk: ``[['[CLS]'*, 'h', 'e', 'l'*, '[SEP]'*],
-            ['[CLS]'*, 'e'*, 'l', 'l'*, '[SEP]'*], ['[CLS]'*, 'l'*, 'l', 'o', '[SEP]'*]]``.
-    """
-
-    @property
-    def output_types(self) -> Optional[Dict[str, NeuralType]]:
-        """Returns definitions of module output ports.
-
-        input_ids: ids of word subtokens encoded using tokenizer
-        segment_ids: an array of zeros
-        input_mask: attention mask. Zeros if input is padding.
-        subtoken_mask: a mask used for retrieving predictions for words. An element equals ``1`` if corresponding
-            token is the first token in some word and zero otherwise. For example, if input query
-            "language processing" is tokenized into ["[CLS]", "language", "process", "ing", "SEP"], then
-            ``subtokens_mask`` will be [0, 1, 1, 0, 0].
-        quantities_of_preceding_words: number of words preceding a segment in a query. It is used for uniting
-            predictions from different segments if such segments overlap. For example, if query "hello john" is
-            tokenized into segments ``[['hell', 'o'], ['john']]``, then ``quantities_of_preceding_words=[0, 1]``.
-        query_ids: ids of queries to which segments belong. For example, if ``queries=["foo", "bar"]`` are
-            segmented into ``[[['[CLS]', 'f', 'o', '[SEP]'], ['[CLS]', 'o', 'o', '[SEP]']],
-            [['[CLS]', 'b', 'a', '[SEP]'], ['[CLS]', 'a', 'r', '[SEP]']]]``, then for batch
-            [['[CLS]', 'o', 'o', '[SEP]'], ['[CLS]', 'b', 'a', '[SEP]'], ['[CLS]', 'a', 'r', '[SEP]']]
-            ``query_ids=[0, 1, 1]``.
-        is_first: is segment the first segment in query. The left margin of the first segment in a query is not
-            removed and this parameter is used to identify first segments.
-        is_last: is segment the last segment in query. The right margin of the last segment in a query is not
-            removed and this parameter is used to identify last segments.
-
-        """
-        return {
-            'input_ids': NeuralType(('B', 'T'), ChannelType()),
-            'segment_ids': NeuralType(('B', 'T'), ChannelType()),
-            'input_mask': NeuralType(('B', 'T'), MaskType()),
-            'subtokens_mask': NeuralType(('B', 'T'), MaskType()),
-            'quantities_of_preceding_words': NeuralType(('B',), Index()),
-            'query_ids': NeuralType(('B',), Index()),
-            'is_first': NeuralType(('B',), BoolType()),
-            'is_last': NeuralType(('B',), BoolType()),
-        }
-
-    def __init__(
-        self, queries: List[str], tokenizer: TokenizerSpec, max_seq_length: int = 128, step: int = 32, margin: int = 16
-    ):
-        features = get_features_infer(
-            queries=queries, max_seq_length=max_seq_length, tokenizer=tokenizer, step=step, margin=margin
-        )
-        self.all_input_ids: List[List[int]] = features[0]
-        self.all_segment_ids: List[List[int]] = features[1]
-        self.all_input_mask: List[List[int]] = features[2]
-        self.all_subtokens_mask: List[List[int]] = features[3]
-        self.all_quantities_of_preceding_words: List[int] = features[4]
-        self.all_query_ids: List[int] = features[5]
-        self.all_is_first: List[bool] = features[6]
-        self.all_is_last: List[bool] = features[7]
-
-    def __len__(self) -> int:
-        return len(self.all_input_ids)
-
-    def collate_fn(
-        self, batch: List[Tuple[np.nddarray, np.ndarray, np.ndarray, np.ndarray, int, int, bool, bool]]
-    ) -> Tuple[
-        torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, Tuple[int], Tuple[int], Tuple[bool], Tuple[bool]
-    ]:
-        inp_ids, segment_ids, inp_mask, st_mask, n_preceding, query_ids, is_first, is_last = zip(*batch)
-        return (
-            pad_sequence([torch.tensor(x) for x in inp_ids], batch_first=True, padding_value=0),
-            pad_sequence([torch.tensor(x) for x in segment_ids], batch_first=True, padding_value=0),
-            pad_sequence([torch.tensor(x) for x in inp_mask], batch_first=True, padding_value=0),
-            pad_sequence([torch.tensor(x) for x in st_mask], batch_first=True, padding_value=0),
-            n_preceding,
-            query_ids,
-            is_first,
-            is_last,
-        )
-
-    def __getitem__(self, idx: int) -> Tuple[np.nddarray, np.ndarray, np.ndarray, np.ndarray, int, int, bool, bool]:
-        return (
-            np.array(self.all_input_ids[idx]),
-            np.array(self.all_segment_ids[idx]),
-            np.array(self.all_input_mask[idx], dtype=np.float32),
-            np.array(self.all_subtokens_mask[idx]),
-            self.all_quantities_of_preceding_words[idx],
-            self.all_query_ids[idx],
-            self.all_is_first[idx],
-            self.all_is_last[idx],
-        )
-
-
-def setup_infer_long_dataloader(
-    model: PunctuationCapitalizationModel,
-    queries: List[str],
-    batch_size: int,
-    max_seq_length: int,
-    step: int,
-    margin: int,
-) -> torch.utils.data.DataLoader:
-    """
-    Setup function for a infer data loader.
-
-    Args:
-        model: a ``PunctuationCapitalizationModel`` instance for which data loader is created.
-        queries: lower cased text without punctuation
-        batch_size: batch size to use during inference
-        max_seq_length: length of segments into which queries are split. ``max_seq_length`` includes ``[CLS]`` and
-            ``[SEP]`` so every segment contains at most ``max_seq_length-2`` tokens from input a query.
-        step: number of tokens by which a segment is offset to a previous segment. Parameter ``step`` cannot be greater
-            than ``max_seq_length-2``.
-        margin: number of tokens near the edge of a segment which label probabilities are not used in final prediction
-            computation.
-    Returns:
-        A pytorch DataLoader.
-    """
-    if max_seq_length is None:
-        max_seq_length = model._cfg.dataset.max_seq_length
-    if step is None:
-        step = model._cfg.dataset.step
-    if margin is None:
-        margin = model._cfg.dataset.margin
-
-    dataset = BertPunctuationCapitalizationInferLongDataset(
-        tokenizer=model.tokenizer, queries=queries, max_seq_length=max_seq_length, step=step, margin=margin
-    )
-    return torch.utils.data.DataLoader(
-        dataset=dataset,
-        collate_fn=dataset.collate_fn,
-        batch_size=batch_size,
-        shuffle=False,
-        num_workers=model._cfg.dataset.num_workers,
-        pin_memory=model._cfg.dataset.pin_memory,
-        drop_last=False,
-    )
-
-
-def _move_acc_probs_to_token_preds(
-    pred: List[int], acc_prob: np.ndarray, number_of_probs_to_move: int
-) -> Tuple[List[int], np.ndarray]:
-    """
-    ``number_of_probs_to_move`` rows in the beginning are removed from ``acc_prob``. From every remove row the label
-    with the largest probability is selected and appended to ``pred``.
-    Args:
-        pred: list with ready label indices for a query
-        acc_prob: numpy array of shape ``[number_of_words_for_which_probabilities_are_accumulated, number_of_labels]``
-        number_of_probs_to_move: int
-    Returns:
-        pred: list with ready label indices for a query
-        acc_prob: numpy array of shape
-            ``[number_of_words_for_which_probabilities_are_accumulated - number_of_probs_to_move, number_of_labels]``
-    """
-    if number_of_probs_to_move > acc_prob.shape[0]:
-        raise ValueError(
-            f"Not enough accumulated probabilities. Number_of_probs_to_move={number_of_probs_to_move} "
-            f"acc_prob.shape={acc_prob.shape}"
-        )
-    if number_of_probs_to_move > 0:
-        pred = pred + list(np.argmax(acc_prob[:number_of_probs_to_move], axis=-1))
-    acc_prob = acc_prob[number_of_probs_to_move:]
-    return pred, acc_prob
-
-
-def _update_accumulated_probabilities(acc_prob: np.ndarray, update: np.ndarray) -> np.ndarray:
-    """
-    Args:
-        acc_prob: numpy array of shape ``[A, L]``
-        update: numpy array of shape ``[A + N, L]``
-    Returns:
-        numpy array of shape ``[A + N, L]``
-    """
-    acc_prob *= update[: acc_prob.shape[0]]
-    acc_prob = np.concatenate([acc_prob, update[acc_prob.shape[0] :]], axis=0)
-    return acc_prob
-
-
-def _remove_margins(tensor, margin_size, keep_left, keep_right):
-    if not keep_left:
-        tensor = tensor[margin_size + 1 :]  # remove left margin and CLS token
-    if not keep_right:
-        tensor = tensor[: tensor.shape[0] - margin_size - 1]  # remove right margin and SEP token
-    return tensor
-
-
-def apply_punct_capit_predictions(cfg: DictConfig, query: str, punct_preds: List[int], capit_preds: List[int]) -> str:
-    """
-    Restores punctuation and capitalization in ``query``.
-    Args:
-        cfg: config for ``nemo.collections.nlp.models.PunctuationCapitalizationModel``
-        query: a string without punctuation and capitalization
-        punct_preds: ids of predicted punctuation labels
-        capit_preds: ids of predicted capitalization labels
-    Returns:
-        a query with restored punctuation and capitalization
-    """
-    query = query.strip().split()
-    assert len(query) == len(
-        punct_preds
-    ), f"len(query)={len(query)} len(punct_preds)={len(punct_preds)}, query[:30]={query[:30]}"
-    assert len(query) == len(
-        capit_preds
-    ), f"len(query)={len(query)} len(capit_preds)={len(capit_preds)}, query[:30]={query[:30]}"
-    punct_ids_to_labels = {v: k for k, v in cfg.punct_label_ids.items()}
-    capit_ids_to_labels = {v: k for k, v in cfg.capit_label_ids.items()}
-    query_with_punct_and_capit = ''
-    for j, word in enumerate(query):
-        punct_label = punct_ids_to_labels[punct_preds[j]]
-        capit_label = capit_ids_to_labels[capit_preds[j]]
-
-        if capit_label != cfg.dataset.pad_label:
-            word = word.capitalize()
-        query_with_punct_and_capit += word
-        if punct_label != cfg.dataset.pad_label:
-            query_with_punct_and_capit += punct_label
-        query_with_punct_and_capit += ' '
-    return query_with_punct_and_capit[:-1]
-
-
-def _transform_logit_to_prob_and_remove_margins_and_extract_word_probs(
-    punct_logits: torch.Tensor,
-    capit_logits: torch.Tensor,
-    subtokens_mask: torch.Tensor,
-    start_word_ids: Tuple[int],
-    margin: int,
-    is_first: Tuple[bool],
-    is_last: Tuple[bool],
-) -> Tuple[List[np.ndarray], List[np.ndarray], List[int]]:
-    """
-    Applies softmax to get punctuation and capitalization probabilities, applies ``subtokens_mask`` to extract
-    probabilities for words from probabilities for tokens, removes ``margin`` probabilities near edges of a segment.
-    Left margin of the first segment in a query and right margin of the last segment in a query are not removed.
-    Calculates new ``start_word_ids`` taking into the account the margins. If the left margin of a segment is removed
-    corresponding start word index is increased by number of words (number of nonzero values in corresponding
-    ``subtokens_mask``) in the margin.
-    Args:
-        punct_logits: a float tensor of shape ``[batch_size, segment_length, number_of_punctuation_labels]``
-        capit_logits: a float tensor of shape ``[batch_size, segment_length, number_of_capitalization_labels]``
-        subtokens_mask: a float tensor of shape ``[batch_size, segment_length]``
-        start_word_ids: indices of segment first words in a query
-        margin: number of tokens near edges of a segment which probabilities are discarded
-        is_first: is segment the first segment in a query
-        is_last: is segment the last segment in a query
-    Returns:
-        b_punct_probs: list containing ``batch_size`` numpy arrays. The numpy arrays have shapes
-            ``[number_of_word_in_this_segment, number_of_punctuation_labels]``. Word punctuation probabilities for
-            segments in the batch.
-        b_capit_probs: list containing ``batch_size`` numpy arrays. The numpy arrays have shapes
-            ``[number_of_word_in_this_segment, number_of_capitalization_labels]``. Word capitalization probabilities for
-            segments in the batch.
-        new_start_word_ids: indices of segment first words in a query after margin removal
-    """
-    new_start_word_ids = list(start_word_ids)
-    subtokens_mask = subtokens_mask > 0.5
-    b_punct_probs, b_capit_probs = [], []
-    for i, (first, last, pl, cl, stm) in enumerate(zip(is_first, is_last, punct_logits, capit_logits, subtokens_mask)):
-        if not first:
-            new_start_word_ids[i] += torch.count_nonzero(stm[: margin + 1]).numpy()  # + 1 is for [CLS] token
-        stm = _remove_margins(stm, margin, keep_left=first, keep_right=last)
-        for b_probs, logits in [(b_punct_probs, pl), (b_capit_probs, cl)]:
-            p = torch.nn.functional.softmax(
-                _remove_margins(logits, margin, keep_left=first, keep_right=last)[stm], dim=-1,
-            )
-            b_probs.append(p.detach().cpu().numpy())
-    return b_punct_probs, b_capit_probs, new_start_word_ids
-
-
-def add_punctuation_capitalization(
-    model: PunctuationCapitalizationModel,
-    queries: List[str],
-    batch_size: int = None,
-    max_seq_length: int = 64,
-    step: int = 8,
-    margin: int = 16,
-) -> List[str]:
-    """
-    Adds punctuation and capitalization to the queries. Use this method for inference.
-
-    Parameters ``max_seq_length``, ``step``, ``margin`` are for controlling the way queries are split into segments
-    which then processed by the model. Parameter ``max_seq_length`` is a length of a segment after tokenization
-    including special tokens [CLS] in the beginning and [SEP] in the end of a segment. Parameter ``step`` is shift
-    between consequent segments. Parameter ``margin`` is used to exclude negative effect of subtokens near
-    borders of segments which have only one side context.
-
-    If segments overlap, probabilities of overlapping predictions are multiplied and then the label with
-    corresponding to the maximum probability is selected.
-
-    Args:
-        queries: lower cased text without punctuation
-        batch_size: batch size to use during inference
-        max_seq_length: maximum sequence length of segment after tokenization.
-        step: relative shift of consequent segments into which long queries are split. Long queries are split into
-            segments which can overlap. Parameter ``step`` controls such overlapping. Imagine that queries are
-            tokenized into characters, ``max_seq_length=5``, and ``step=2``. In such a case query "hello" is
-            tokenized into segments ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'l', 'l', 'o', '[SEP]']]``.
-        margin: number of subtokens in the beginning and the end of segments which are not used for prediction
-            computation. The first segment does not have left margin and the last segment does not have right
-            margin. For example, if input sequence is tokenized into characters, ``max_seq_length=5``,
-            ``step=1``, and ``margin=1``, then query "hello" will be tokenized into segments
-            ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'e', 'l', 'l', '[SEP]'],
-            ['[CLS]', 'l', 'l', 'o', '[SEP]']]``. These segments are passed to the model. Before final predictions
-            computation, margins are removed. In the next list, subtokens which logits are not used for final
-            predictions computation are marked with asterisk: ``[['[CLS]'*, 'h', 'e', 'l'*, '[SEP]'*],
-            ['[CLS]'*, 'e'*, 'l', 'l'*, '[SEP]'*], ['[CLS]'*, 'l'*, 'l', 'o', '[SEP]'*]]``.
-    Returns:
-        result: text with added capitalization and punctuation ``max_seq_length`` equals 5, ``step`` equals 2, and
-
-    """
-    if len(queries) == 0:
-        return []
-    if batch_size is None:
-        batch_size = len(queries)
-        logging.info(f'Using batch size {batch_size} for inference')
-    result: List[str] = []
-    mode = model.training
-    d = 'cuda' if torch.cuda.is_available() else 'cpu'
-    try:
-        model.eval()
-        model = model.to(d)
-        infer_datalayer = setup_infer_long_dataloader(model, queries, batch_size, max_seq_length, step, margin)
-        # Predicted labels for queries. List of labels for every query
-        all_punct_preds: List[List[int]] = [[] for _ in queries]
-        all_capit_preds: List[List[int]] = [[] for _ in queries]
-        # Accumulated probabilities (or product of probabilities acquired from different segments) of punctuation
-        # and capitalization. Probabilities for words in a query are extracted using `subtokens_mask`. Probabilities
-        # for newly processed words are appended to the accumulated probabilities. If probabilities for a word are
-        # already present in `acc_probs`, old probabilities are replaced with a product of old probabilities
-        # and probabilities acquired from new segment. Segments are processed in an order they appear in an
-        # input query. When all segments with a word are processed, a label with the highest probability
-        # (or product of probabilities) is chosen and appended to an appropriate list in `all_preds`. After adding
-        # prediction to `all_preds`, probabilities for a word are removed from `acc_probs`.
-        acc_punct_probs: List[Optional[np.ndarray]] = [None for _ in queries]
-        acc_capit_probs: List[Optional[np.ndarray]] = [None for _ in queries]
-        for batch_i, batch in enumerate(infer_datalayer):
-            inp_ids, inp_type_ids, inp_mask, subtokens_mask, start_word_ids, query_ids, is_first, is_last = batch
-            punct_logits, capit_logits = model.forward(
-                input_ids=inp_ids.to(d), token_type_ids=inp_type_ids.to(d), attention_mask=inp_mask.to(d),
-            )
-            _res = _transform_logit_to_prob_and_remove_margins_and_extract_word_probs(
-                punct_logits, capit_logits, subtokens_mask, start_word_ids, margin, is_first, is_last
-            )
-            punct_probs, capit_probs, start_word_ids = _res
-            for i, (q_i, start_word_id, bpp_i, bcp_i) in enumerate(
-                zip(query_ids, start_word_ids, punct_probs, capit_probs)
-            ):
-                for all_preds, acc_probs, b_probs_i in [
-                    (all_punct_preds, acc_punct_probs, bpp_i),
-                    (all_capit_preds, acc_capit_probs, bcp_i),
-                ]:
-                    if acc_probs[q_i] is None:
-                        acc_probs[q_i] = b_probs_i
-                    else:
-                        all_preds[q_i], acc_probs[q_i] = _move_acc_probs_to_token_preds(
-                            all_preds[q_i], acc_probs[q_i], start_word_id - len(all_preds[q_i]),
-                        )
-                        acc_probs[q_i] = _update_accumulated_probabilities(acc_probs[q_i], b_probs_i)
-        for all_preds, acc_probs in [(all_punct_preds, acc_punct_probs), (all_capit_preds, acc_capit_probs)]:
-            for q_i, (pred, prob) in enumerate(zip(all_preds, acc_probs)):
-                if prob is not None:
-                    all_preds[q_i], acc_probs[q_i] = _move_acc_probs_to_token_preds(pred, prob, len(prob))
-        for i, query in enumerate(queries):
-            result.append(apply_punct_capit_predictions(model._cfg, query, all_punct_preds[i], all_capit_preds[i]))
-    finally:
-        # set mode back to its original value
-        model.train(mode=mode)
-    return result
-
-
 def main():
     args = get_args()
     if args.pretrained_name is None:
@@ -716,8 +173,7 @@ def main():
         texts = []
         for item in manifest:
             texts.append(item[text_key])
-    processed_texts = add_punctuation_capitalization(
-        model,
+    processed_texts = model.add_punctuation_capitalization(
         texts,
         batch_size=args.batch_size,
         max_seq_length=args.max_seq_length,
